Revert "JP-1413 NIRSpec IFU level-3 ASNs to segregate obs based on the grating (#4867)"
This reverts commit 1785cbc33d11c36cb8929ce2c4304bcdc0420e39.
